schedule/requirements.txt
mc-convention>=1.0
Django>=1.11
icalendar
